MapAndRoomNameSelector script and more
- Created MapAndRoomNameSelector.cs for map and room name selection when creating custom game. (still in progress)
- Moved CustomBattleGameModeSelector.cs under CreateRoom folder and namespace.
- Added reference to MenuUi.Scripts.Lobby.CreateRoom namespace in BattlePopupPanelManager.cs since it holds the CustomGameMode enum now.
- Updated CreateRoomCustom.cs to access map and room name selector and custom battle game mode selector. Getting room name selected emotion and selected custom game mode.
- Connected references in Battle Popup prefab.
