+++ Assets/MenuUi/Scripts/Lobby/CreateRoom/MapAndRoomNameSelector.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: babf3a92087364943a445bdf67a9572e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
